Refactor input streaming into separate module
